解决以下问题：1.使用Redis 配置替换fastjson 反序列化报错 com.alibaba.fastjson.JSONException: autoType is not support 添加autotype白名单 2.引入Jedis 报错：java.lang.ClassNotFoundException: redis.clients.util.SafeEncoder 2.9.0版本一直报这个错,然后将版本升级到3.1.0解决问题

+++ src/main/java/com/example/aop/redis/demo/redis/EnhanceRedisTemplate.java
@@ -23,6 +23,13 @@ public class EnhanceRedisTemplate extends RedisTemplate {
         }
     };
 
+    /**
+     * 重写RedisConnection方法,设置Redis库的编号
+     *
+     * @param connection
+     * @param existingConnection
+     * @return
+     */
     @Override
     protected RedisConnection preProcessConnection(RedisConnection connection, boolean existingConnection) {
         try {
@@ -30,7 +37,7 @@ public class EnhanceRedisTemplate extends RedisTemplate {
             //如果设置了dbIndex
             if (dbIndex != null) {
                 if (connection instanceof JedisConnection) {
-                    if (((JedisConnection) connection).getNativeConnection().getDB().intValue() != dbIndex) {
+                    if (((JedisConnection) connection).getNativeConnection().getDB() != dbIndex) {
                         connection.select(dbIndex);
                     }
                 } else {
